Refactor directory structure again.
The directory layout has changed again (sorry for that). It happens, that folks only work on a single Vagrantfile untit we find the final mixture. Maintaining a second Vagrantfile is simply overkill at the moment. For now, you will find the latest and greatest Vagrant setup directly in the directories "openshift", "openshift-cluster" and "kubernetes".