Extended Vagrantfile to permit use of Proxmox provider.
The libvirt box can be used to manually create a VM on a Proxmox cluster that can then subsequently be cloned by Vagrant.

In order for this to work, it is necessary to be able to set a static IP address for the clone with a predefined MAC address.

It is also desirable to disable Landrush or Hostmanager mapping of name lookups and to use local network DNS.